backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
alembic>=1.14.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
redis>=5.0.0
celery>=5.4.0
boto3>=1.35.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
transformers>=4.45.0
torch>=2.0.0
open-clip-torch>=2.26.0
numpy>=1.26.0
Pillow>=10.0.0
python-slugify>=8.0.0
httpx>=0.27.0
prometheus-client>=0.20.0
lingua-language-detector>=2.0.0
bm25s>=0.2.0
PyStemmer>=2.2.0
